Created useMyUserData hook for handling user data from DB
* Created hooks folder for storing all hooks
* Refactor away from directly using codegen queries
