+++ src/pages/create.tsx
@@ -4,11 +4,10 @@ import { Button, Input, Text } from "../components/atomic";
 import {
   CreateProgramInput,
   useCreateProgramMutation,
-  useGetMyUserQuery,
 } from "../generated/graphql";
+import { useMyUserData } from "../hooks";
 import RedirectIfNotLoggedIn from "../layouts/RedirectIfNotLoggedIn";
 import Page from "../types/Page";
-import { useAuth } from "../utils/firebase/auth";
 
 const BlobCircle = () => {
   const sizes = "h-24 w-24 md:h-64 md:w-64 lg:h-80 lg:w-80";
@@ -31,8 +30,7 @@ const CreateProgramPage: Page = () => {
   const [displayError, setError] = useState(""); // TODO: Proper error box
   const [loading, setLoading] = useState(false);
   const router = useRouter();
-  const { user } = useAuth();
-  const { refetch: refetchMyUser } = useGetMyUserQuery({ skip: !user });
+  const { refetchMyUserData } = useMyUserData();
 
   const validateProgramName = (name: string) => {
     // check string is not whitespace; can ask product for more constraints on names
@@ -59,7 +57,7 @@ const CreateProgramPage: Page = () => {
     return createProgram({ variables: { data: createProgramInput } })
       .then(async (res) => {
         if (res) {
-          if (refetchMyUser) await refetchMyUser();
+          if (refetchMyUserData) await refetchMyUserData();
           router.push(`/program/${res.data?.createProgram.slug}`);
 
           // If we use subdomains
